+++ login.html
@@ -68,6 +68,6 @@
         <p>판다마켓이 처음이신가요?<a href="/signup.html">회원가입</a></p>
       </section>
     </main>
-    <script src="./js/login.js"></script>
+    <script type="module" src="./js/login.js"></script>
   </body>
 </html>
